C++11 variadic templates based function_eval.
