+++ include/boost/phoenix/core/detail/function_eval.hpp
@@ -1,55 +1,48 @@
-#if !defined(BOOST_PHOENIX_DONT_USE_PREPROCESSED_FILES)
-
-#ifndef BOOST_PHOENIX_CORE_DETAIL_FUNCTION_EVAL_HPP
-#define BOOST_PHOENIX_CORE_DETAIL_FUNCTION_EVAL_HPP
-
-#include <boost/phoenix/core/limits.hpp>
-#include <boost/phoenix/support/iterate.hpp>
-#include <boost/phoenix/core/call.hpp>
-#include <boost/phoenix/core/expression.hpp>
-#include <boost/phoenix/core/meta_grammar.hpp>
-#include <boost/phoenix/core/detail/phx2_result.hpp>
-#include <boost/utility/result_of.hpp>
-
-#include <boost/phoenix/core/detail/preprocessed/function_eval.hpp>
-
-#endif
-#else
-
-#if !BOOST_PHOENIX_IS_ITERATING
-
-#ifndef BOOST_PHOENIX_CORE_DETAIL_FUNCTION_EVAL_HPP
-#define BOOST_PHOENIX_CORE_DETAIL_FUNCTION_EVAL_HPP
-
-#include <boost/phoenix/core/limits.hpp>
-#include <boost/phoenix/support/iterate.hpp>
-#include <boost/phoenix/core/call.hpp>
-#include <boost/phoenix/core/expression.hpp>
-#include <boost/phoenix/core/meta_grammar.hpp>
-#include <boost/phoenix/core/detail/phx2_result.hpp>
-#include <boost/utility/result_of.hpp>
-
-#if defined(__WAVE__) && defined(BOOST_PHOENIX_CREATE_PREPROCESSED_FILES)
-#pragma wave option(preserve: 2, line: 0, output: "preprocessed/function_eval_" BOOST_PHOENIX_LIMIT_STR ".hpp")
-#endif
 /*=============================================================================
     Copyright (c) 2001-2007 Joel de Guzman
+    Copyright (c) 2015 Kohei Takahashi
 
     Distributed under the Boost Software License, Version 1.0. (See accompanying
     file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 ==============================================================================*/
+#ifndef BOOST_PHOENIX_CORE_DETAIL_FUNCTION_EVAL_HPP
+#define BOOST_PHOENIX_CORE_DETAIL_FUNCTION_EVAL_HPP
 
+#include <boost/config.hpp>
+#include <boost/phoenix/support/iterate.hpp>
+#include <boost/phoenix/core/call.hpp>
+#include <boost/phoenix/core/expression.hpp>
+#include <boost/phoenix/core/meta_grammar.hpp>
+#include <boost/phoenix/core/detail/phx2_result.hpp>
+#include <boost/utility/result_of.hpp>
 
-#if defined(__WAVE__) && defined(BOOST_PHOENIX_CREATE_PREPROCESSED_FILES)
-#pragma wave option(preserve: 1)
+#if defined(BOOST_NO_CXX11_VARIADIC_TEMPLATES) || \
+     defined(BOOST_NO_CXX11_RVALUE_REFERENCES)
+#   include <boost/phoenix/core/limits.hpp>
+#   define BOOST_PHOENIX_NO_VARIADIC_FUNCTION_EVAL
+#else
+#   include <boost/mpl/if.hpp>
+#   include <boost/type_traits/is_reference.hpp>
 #endif
 
+#if !defined(BOOST_PHOENIX_DONT_USE_PREPROCESSED_FILES)
+#   include <boost/phoenix/core/detail/preprocessed/function_eval_expression.hpp>
+#else
+#if defined(__WAVE__) && defined(BOOST_PHOENIX_CREATE_PREPROCESSED_FILES)
+#pragma wave option(preserve: 2, line: 0, output: "preprocessed/function_eval_expression_" BOOST_PHOENIX_LIMIT_STR ".hpp")
+#endif
+// XXX: Currently, following definition should be preprocessed since
+// core/expression.hpp does not support c++11 variadic templates yet.
 BOOST_PHOENIX_DEFINE_EXPRESSION_VARARG(
     (boost)(phoenix)(detail)(function_eval)
   , (meta_grammar)
     (meta_grammar)
   , BOOST_PP_DEC(BOOST_PHOENIX_COMPOSITE_LIMIT)
 )
+#if defined(__WAVE__) && defined(BOOST_PHOENIX_CREATE_PREPROCESSED_FILES)
+#pragma wave option(output: null)
+#endif
+#endif
 
 namespace boost { namespace phoenix {
     namespace detail
@@ -59,7 +52,7 @@ namespace boost { namespace phoenix {
         {
             return x;
         }
-        
+
         template <typename T>
         T const& help_rvalue_deduction(T const& x)
         {
@@ -71,6 +64,7 @@ namespace boost { namespace phoenix {
             template <typename Sig>
             struct result;
 
+#ifndef BOOST_PHOENIX_NO_VARIADIC_FUNCTION_EVAL
             template <typename This, typename F, typename Context>
             struct result<This(F, Context)>
             {
@@ -97,37 +91,71 @@ namespace boost { namespace phoenix {
                 return boost::phoenix::eval(f, ctx)();
             }
 
-        #define PHOENIX_GET_ARG(z, n, data)                                     \
-            typedef                                                             \
-                typename boost::add_reference<                                  \
-                    typename boost::add_const<                                  \
-                        typename boost::result_of<                              \
-                            boost::phoenix::evaluator(                          \
-                                BOOST_PP_CAT(A, n)                              \
-                              , Context                                         \
-                            )                                                   \
-                        >::type                                                 \
-                    >::type                                                     \
-                >::type                                                         \
-                BOOST_PP_CAT(a, n);
+            #include <boost/phoenix/core/detail/cpp03/function_eval.hpp>
+#else
+            template <typename, typename, typename...> struct result_impl;
 
-        #define PHOENIX_EVAL_ARG(z, n, data)                                    \
-            help_rvalue_deduction(boost::phoenix::eval(BOOST_PP_CAT(a, n), ctx))
-        
-        #define M0(z, n, data)                                     \
-            typename proto::detail::uncvref<BOOST_PP_CAT(a, n)>::type
+            template <typename F, typename... A, typename Head, typename... Tail>
+            struct result_impl<F, void(A...), Head, Tail...>
+                : result_impl<F, void(A..., Head), Tail...>
+            {
+            };
 
-        #define BOOST_PHOENIX_ITERATION_PARAMS                                  \
-            (3, (1, BOOST_PP_DEC(BOOST_PHOENIX_ACTOR_LIMIT),                    \
-            <boost/phoenix/core/detail/function_eval.hpp>))
-#include BOOST_PHOENIX_ITERATE()
+            template <typename F, typename... A, typename Context>
+            struct result_impl<F, void(A...), Context>
+            {
+                typedef typename
+                    remove_reference<
+                        typename boost::result_of<evaluator(F, Context)>::type
+                    >::type
+                    fn;
 
-        #undef PHOENIX_GET_ARG
-        #undef PHOENIX_EVAL_ARG
-        #undef M0
-        
+
+                typedef typename
+                    boost::result_of<
+                        fn(
+                            typename boost::add_reference<
+                                typename boost::add_const<
+                                    typename boost::result_of<boost::phoenix::evaluator(A, Context)>::type
+                                >::type
+                            >::type...
+                        )
+                    >::type
+                    type;
+
+                static type call(F f, A... a, Context ctx)
+                {
+                    return boost::phoenix::eval(f, ctx)(help_rvalue_deduction(boost::phoenix::eval(a, ctx))...);
+                }
+            };
+
+            template <typename This, typename F, typename... A>
+            struct result<This(F, A...)>
+                : result_impl<F, void(), A...>
+            {
+            };
+
+            template <typename F, typename... A>
+            typename result<
+                function_eval(
+                    F const &
+                  , typename mpl::if_<is_reference<A>, A, A const &>::type...
+                )
+            >::type
+            // 'A &... a, Context const &ctx' doesn't work as intended: type deduction always fail.
+            operator()(F && f, A &&... a) const
+            {
+                return
+                    result<
+                        function_eval(
+                            typename mpl::if_<is_reference<F>, F, F const &>::type
+                          , typename mpl::if_<is_reference<A>, A, A const &>::type...
+                        )
+                    >::call(f, a...);
+            }
+#endif
         };
-        
+
     }
 
     template <typename Dummy>
@@ -136,79 +164,5 @@ namespace boost { namespace phoenix {
     {};
 }}
 
-#if defined(__WAVE__) && defined(BOOST_PHOENIX_CREATE_PREPROCESSED_FILES)
-#pragma wave option(output: null)
 #endif
 
-
-#endif
-
-#else
-            template <
-                typename This
-              , typename F
-              , BOOST_PHOENIX_typename_A
-              , typename Context
-            >
-            struct result<This(F, BOOST_PHOENIX_A, Context)>
-            {
-                typedef typename
-                    remove_reference<
-                        typename boost::result_of<evaluator(F, Context)>::type
-                    >::type
-                    fn;
-
-                BOOST_PP_REPEAT(BOOST_PHOENIX_ITERATION, PHOENIX_GET_ARG, _)
-
-                typedef typename
-                    boost::result_of<fn(BOOST_PHOENIX_a)>::type
-                    type;
-                /*
-                typedef typename
-                    mpl::eval_if_c<
-                        has_phx2_result<
-                            fn
-                          , BOOST_PP_ENUM(BOOST_PHOENIX_ITERATION, M0, _)
-                        >::value
-                      , boost::result_of<
-                            fn(
-                                BOOST_PHOENIX_a
-                            )
-                        >
-                      , phx2_result<
-                            fn
-                          , BOOST_PHOENIX_a
-                        >
-                    >::type
-                    type;
-                */
-            };
-
-            template <typename F, BOOST_PHOENIX_typename_A, typename Context>
-            typename result<
-                function_eval(
-                    F const &
-                  , BOOST_PHOENIX_A_ref
-                  , Context const &
-                )
-            >::type
-            operator()(F const & f, BOOST_PHOENIX_A_ref_a, Context const & ctx) const
-            {
-                return boost::phoenix::eval(f, ctx)(BOOST_PP_ENUM(BOOST_PHOENIX_ITERATION, PHOENIX_EVAL_ARG, _));
-            }
-
-            template <typename F, BOOST_PHOENIX_typename_A, typename Context>
-            typename result<
-                function_eval(
-                    F &
-                  , BOOST_PHOENIX_A_ref
-                  , Context const &
-                )
-            >::type
-            operator()(F & f, BOOST_PHOENIX_A_ref_a, Context const & ctx) const
-            {
-                return boost::phoenix::eval(f, ctx)(BOOST_PP_ENUM(BOOST_PHOENIX_ITERATION, PHOENIX_EVAL_ARG, _));
-            }
-#endif
-
-#endif
